requirements.txt
blessed==1.14.2
boto3==1.4.8
botocore==1.8.38
cachetools==3.1.0
cement==2.8.2
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.7
coverage==4.5.1
decorator==4.1.2
dj-database-url==0.4.1
Django==1.11.18
django-split-settings==0.2.5
django-storages==1.6.6
djangorestframework==3.6.4
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
geocoder==1.32.0
google-api-core==1.7.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
gunicorn==19.6.0
idna==2.6
jmespath==0.9.3
pathspec==0.5.0
protobuf==3.6.1
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.4
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
ratelim==0.1.6
requests==2.21.0
rsa==3.4.2
s3transfer==0.1.12
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
termcolor==1.1.0
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.46.0
whitenoise==3.2
